add Reader as wrapper over BytesReader
